DTO cleanup

+++ src/main/java/io/contek/invoker/binancespot/api/websocket/market/BookTickerEvent.java
@@ -3,14 +3,15 @@ package io.contek.invoker.binancespot.api.websocket.market;
 import io.contek.invoker.binancespot.api.websocket.common.WebSocketEventData;
 
 import javax.annotation.concurrent.NotThreadSafe;
+import java.math.BigDecimal;
 
 @NotThreadSafe
 public class BookTickerEvent extends WebSocketEventData {
 
-  public Long u; // order book updateId
+  public long u; // order book updateId
   public String s; // symbol
-  public Double b; // best bid price
-  public Double B; // best bid qty
-  public Double a; // best ask price
-  public Double A; // best ask qty
+  public BigDecimal b; // best bid price
+  public BigDecimal B; // best bid qty
+  public BigDecimal a; // best ask price
+  public BigDecimal A; // best ask qty
 }
